+++ tests/contrib/hooks/test_spark_submit_hook.py
@@ -706,7 +706,7 @@ class TestSparkSubmitHook(unittest.TestCase):
         self.assertEqual(kill_cmd[3], '--kill')
         self.assertEqual(kill_cmd[4], 'driver-20171128111415-0001')
 
-    @patch('airflow.contrib.kubernetes.kube_client.get_kube_client')
+    @patch('airflow.kubernetes.kube_client.get_kube_client')
     @patch('airflow.contrib.hooks.spark_submit_hook.subprocess.Popen')
     def test_k8s_process_on_kill(self, mock_popen, mock_client_method):
         # Given
