+++ tests/kubernetes/kubernetes_request_factory/__init__.py
@@ -16,4 +16,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-#
