+++ tests/kubernetes/kubernetes_request_factory/test_kubernetes_request_factory.py
@@ -15,11 +15,10 @@
 # specific language governing permissions and limitations
 # under the License.
 
-from airflow.contrib.kubernetes.kubernetes_request_factory.\
-    kubernetes_request_factory import KubernetesRequestFactory
-from airflow.contrib.kubernetes.pod import Pod, Resources
-from airflow.contrib.kubernetes.secret import Secret
-from airflow.contrib.kubernetes.pod_runtime_info_env import PodRuntimeInfoEnv
+from airflow.kubernetes.kubernetes_request_factory.kubernetes_request_factory import KubernetesRequestFactory
+from airflow.kubernetes.pod import Pod, Resources
+from airflow.kubernetes.pod_runtime_info_env import PodRuntimeInfoEnv
+from airflow.kubernetes.secret import Secret
 from parameterized import parameterized
 import unittest
 import copy
